@@ -16,11 +16,13 @@
     "bootstrap-webpack": "0.0.5",
     "classnames": "^2.2.5",
     "font-awesome": "^4.6.3",
-    "react": "15.2.0",
-    "react-dom": "15.2.0",
+    "react": "15.3.0",
+    "react-addons-pure-render-mixin": "^15.3.0",
+    "react-dom": "15.3.0",
     "react-redux": "^4.4.5",
     "react-router": "^2.5.2",
     "react-router-redux": "^4.0.5",
+    "recharts": "^0.12.8",
     "redux": "^3.5.2",
     "redux-immutable": "^3.0.6",
     "redux-module-builder": "0.0.12-0",
@@ -69,7 +71,7 @@
     "phantomjs-prebuilt": "^2.1.7",
     "postcss-loader": "^0.9.1",
     "precss": "^1.4.0",
-    "react-addons-test-utils": "15.2.0",
+    "react-addons-test-utils": "15.3.0",
     "redux-devtools": "^3.3.1",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.0.11",
