@@ -1,30 +1,35 @@
 import {LineChart, Line, XAxis, YAxis, CartesianGrid, Tooltip, Legend} from 'Recharts';
-import React from 'React';
+import React from 'react';
+import _ from 'lodash';
+import PureRenderMixin from 'react-addons-pure-render-mixin';
 
-const data = [
-    {name: 'Page A', uv: 4000, pv: 2400, amt: 2400},
-    {name: 'Page B', uv: 3000, pv: 1398, amt: 2210},
-    {name: 'Page C', uv: 2000, pv: 9800, amt: 2290},
-    {name: 'Page D', uv: 2780, pv: 3908, amt: 2000},
-    {name: 'Page E', uv: 1890, pv: 4800, amt: 2181},
-    {name: 'Page F', uv: 2390, pv: 3800, amt: 2500},
-    {name: 'Page G', uv: 3490, pv: 4300, amt: 2100},
-];
-const SimpleLineChart = React.createClass({
+class SimpleLineChart extends React.Component {
+
+    constructor(props) {
+        super(props);
+        this.shouldComponentUpdate = PureRenderMixin.shouldComponentUpdate.bind(this);
+    }
+
     render () {
+
+        const data = _.map(this.props.data.toJS(), function(test,i){
+            return { name:i.toString(), correct:test.correct.length };
+        });
+
+        while (data.length < 10) {
+            data.push({ name:data.length });
+        }
+
         return (
-            <LineChart width={600} height={300} data={data}
-                       margin={{top: 5, right: 30, left: 20, bottom: 5}}>
-                <XAxis dataKey="name"/>
-                <YAxis/>
-                <CartesianGrid strokeDasharray="3 3"/>
-                <Tooltip/>
-                <Legend />
-                <Line type="monotone" dataKey="pv" stroke="#8884d8" activeDot={{r: 8}}/>
-                <Line type="monotone" dataKey="uv" stroke="#82ca9d" />
+            <LineChart width={600} height={300} data={data}>
+                <YAxis domain={[0, 'dataMax + 10']} />
+                <XAxis dataKey="name" padding={{left: 5}} />
+                <CartesianGrid strokeDasharray="3 3"  />
+                <Line type="monotone" dataKey="correct"  stroke="orange" />
+
             </LineChart>
         );
     }
-})
+}
 
 export default SimpleLineChart;
